src/api/youtube.py
from googleapiclient.discovery import build
from typing import Dict, Any
import logging

class YouTubeAPI:
    def __init__(self, api_key: str):
        self.logger = logging.getLogger(__name__)
        self.youtube = build('youtube', 'v3', developerKey=api_key)
        
    def download_video(self, url: str) -> Dict[str, Any]:
        """Descarga video de YouTube"""
        try:
            video_id = self._extract_video_id(url)
            # Implementar descarga
            return {
                'video_path': f'temp/{video_id}.mp4',
                'metadata': self._get_video_metadata(video_id)
            }
        except Exception as e:
            self.logger.error(f"Error descargando video: {str(e)}")
            raise