requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
python-dotenv>=0.19.0
pyyaml>=5.4.1
pathlib>=1.0.1
fastapi>=0.68.0
uvicorn>=0.15.0
dataclasses>=0.6

# Database
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.1

# Testing
pytest>=7.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.20.0
pytest-cov>=4.0.0

# Video and Image Processing
opencv-python>=4.5.0
moviepy>=1.0.3
scikit-video>=1.1.11
ffmpeg-python>=0.2.0
pytube>=12.1.0

# Machine Learning
torch>=2.0.0
transformers>=4.30.0
ultralytics>=8.0.0
stable-ts>=2.0.0

# Audio Processing
librosa>=0.9.0
soundfile>=0.10.3
pyttsx3>=2.90
whisper>=1.0.0
pytube>=12.1.0
pydub>=0.25.1
ffmpeg-python>=0.2.0

# Google Cloud & AI
google-cloud-aiplatform>=1.35.0
vertexai>=0.0.1
google-cloud-storage>=2.10.0

# Development Tools
black>=22.3.0  # Updated to newer version
flake8>=5.0.0  # Updated to newer version
mypy>=1.0.0    # Updated to newer version
pre-commit>=2.20.0
isort>=5.9.3

# Utils
python-multipart>=0.0.5
requests>=2.26.0
stable-ts>=2.0.0